Corrige 8 bugs encontrados em code review
Críticos (inferência silenciosamente incorreta):
- serve.py, spark_udf.py: BPETokenizer.load() é @classmethod — retorno
  estava sendo descartado, tokenizer ficava sem merges

Alto (corrupção de resultado ou crash):
- tokenizer.py encode(): fallback _merge_rank.get(p, 0) → float("inf")
  para evitar que pares ausentes vençam merges legítimos de rank 0
- tokenizer.py train(): reset self.merges antes de treinar para evitar
  divergência com _merge_rank ao chamar train() mais de uma vez
- dataset.py __len__: max(0, ...) para nunca retornar valor negativo
- trainer.py _eval(): guard contra lista vazia retorna float("nan")
  ao invés de ZeroDivisionError

Médio (consistência e CUDA):
- generate.py, address_eval.py: dtype=torch.long explícito no tensor,
  alinhando com serve.py e spark_udf.py

Baixo (manutenção):
- trainer.py: sincroniza train_cfg.device com self.device resolvido

+++ scripts/address_eval.py
@@ -21,7 +21,7 @@ from slm.tokenizer import BPETokenizer
 def normalizar(model: SLM, tok: BPETokenizer, endereco: str, device: torch.device) -> str:
     prompt = f"ENTRADA: {endereco} | SAIDA:"
     ids = [tok.bos_id] + tok.encode(prompt)
-    idx = torch.tensor([ids], device=device)
+    idx = torch.tensor([ids], dtype=torch.long, device=device)
 
     with torch.no_grad():
         out = model.generate(idx, max_new_tokens=80, temperature=0.3, top_k=20)
